Move the `AsciiHexStream` from `src/core/stream.js` and into its own file
